+++ CHANGELOG.md
@@ -6,6 +6,14 @@ SPDX-License-Identifier: MIT
 
 # Changelog
 
+## tol-sdk `2.7.0`
+
+17-09-2026
+
+- Add quasar() source
+- Add support for direct and via-api versions of datasources
+- Add assembly_span from GoaT
+
 ## tol-sdk `2.6.0`
 
 15-09-2026